+++ main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/leehosu/commitgen/cmd"
+	"github.com/leehosu/commitmate/cmd"
 )
 
 func main() {